requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
streamlit
google-generativeai
python-dotenv
pillow
